Change traversal/merger algorithm
I am relying on a while loop instead of a for loop as it allows us to move directly to one part of the list as I treated the previous files.
I "Close()" the file merger when the appending phase is done.
Removing the "string" import as it is not necessary.
Removing previous algorithm that was commented at the moment.